Put forming code to CounterCard and BigCounterCard
